Barring edit capability to non-logged in users

+++ catalogue.py
@@ -24,7 +24,8 @@ session = DBSession()
 @app.route("/categories/")
 def show_categories():
     categories = session.query(Category).order_by(asc(Category.name))
-    return render_template('categories.html', categories=categories)
+    loggedin = 'username' in login_session
+    return render_template('categories.html', categories=categories, loggedin=loggedin)
 
 
 @app.route("/categories/JSON/")
@@ -36,46 +37,56 @@ def show_categories_json():
 # add a category
 @app.route("/categories/new/", methods=['GET', 'POST'])
 def add_category():
-    if request.method == 'POST':
-        new_category = Category(name=request.form['name'],
-                                description=request.form['description'])
-        session.add(new_category)
-        flash('New category %s added successfully' % new_category.name)
-        session.commit()
-        return redirect(url_for('show_categories'))
+    if 'username' in login_session:
+        if request.method == 'POST':
+            new_category = Category(name=request.form['name'],
+                                    description=request.form['description'])
+            session.add(new_category)
+            flash('New category %s added successfully' % new_category.name)
+            session.commit()
+            return redirect(url_for('show_categories'))
+        else:
+            return render_template('newCategory.html')
     else:
-        return render_template('newCategory.html')
+        return redirect(url_for('login_page'))
 
 
 # edit a category
 @app.route("/categories/<int:category_id>/edit/", methods=['GET', 'POST'])
 def edit_category(category_id):
-    edited_category = session.query(Category).filter_by(id=category_id).one()
-    if request.method == 'POST':
-        if request.form['name'] == edited_category.name and request.form['description'] == edited_category.description:
-            return redirect(url_for('show_categories'))
+    if 'username' in login_session:
+        edited_category = session.query(Category).filter_by(id=category_id).one()
+        if request.method == 'POST':
+            if request.form['name'] == edited_category.name \
+                    and request.form['description'] == edited_category.description:
+                return redirect(url_for('show_categories'))
+            else:
+                if request.form['name']:
+                    edited_category.name = request.form['name']
+                if request.form['description']:
+                    edited_category.description = request.form['description']
+                session.add(edited_category)
+                session.commit()
+                flash('Category %s edited' % edited_category.name)
+                return redirect(url_for('show_categories'))
         else:
-            if request.form['name']:
-                edited_category.name = request.form['name']
-            if request.form['description']:
-                edited_category.description = request.form['description']
-            session.add(edited_category)
-            session.commit()
-            flash('Category %s edited' % edited_category.name)
-            return redirect(url_for('show_categories'))
+            return render_template('editCategory.html', category=edited_category)
     else:
-        return render_template('editCategory.html', category=edited_category)
+        return redirect(url_for('login_page'))
 
 
 # delete a category
 @app.route("/categories/<int:category_id>/delete/", methods=['GET', 'POST'])
 def delete_category(category_id):
-    deleted_category = session.query(Category).filter_by(id=category_id).one()
-    if request.method == 'POST':
-        session.delete(deleted_category)
-        return redirect(url_for('show_categories'))
+    if 'username' in login_session:
+        deleted_category = session.query(Category).filter_by(id=category_id).one()
+        if request.method == 'POST':
+            session.delete(deleted_category)
+            return redirect(url_for('show_categories'))
+        else:
+            return render_template('deleteCategory.html', category=deleted_category)
     else:
-        return render_template('deleteCategory.html', category=deleted_category)
+        return redirect(url_for('login_page'))
 
 
 # show category page and its item list
@@ -83,7 +94,8 @@ def delete_category(category_id):
 def show_category(category_id):
     category = session.query(Category).filter_by(id=category_id).one()
     items = session.query(Item).filter_by(category_id=category_id).order_by(asc(Item.name))
-    return render_template('category.html', category=category, items=items)
+    loggedin = 'username' in login_session
+    return render_template('category.html', category=category, items=items, loggedin=loggedin)
 
 
 @app.route("/categories/<int:category_id>/JSON/")
@@ -96,54 +108,63 @@ def show_categories_json(category_id):
 # add an item
 @app.route("/<int:category_id>/new_item/", methods=['GET', 'POST'])
 def add_item(category_id):
-    if request.method == 'POST':
-        new_item = Item(name=request.form['name'],
-                        description=request.form['description'],
-                        image=request.form['image'],
-                        category_id=category_id)
-        session.add(new_item)
-        flash('New item %s added successfully' % new_item.name)
-        session.commit()
-        return redirect(url_for('show_category', category_id=category_id))
+    if 'username' in login_session:
+        if request.method == 'POST':
+            new_item = Item(name=request.form['name'],
+                            description=request.form['description'],
+                            image=request.form['image'],
+                            category_id=category_id)
+            session.add(new_item)
+            flash('New item %s added successfully' % new_item.name)
+            session.commit()
+            return redirect(url_for('show_category', category_id=category_id))
+        else:
+            return render_template('newItem.html')
     else:
-        return render_template('newItem.html')
+        return redirect(url_for('login_page'))
 
 
 # edit an item
 @app.route("/categories/<int:category_id>/<int:item_id>/edit/", methods=['GET', 'POST'])
 def edit_item(category_id, item_id):
-    category = session.query(Category).filter_by(id=category_id).one()
-    edited_item = session.query(Item).filter_by(id=item_id).one()
-    if request.method == 'POST':
-        if request.form['name'] == edited_item.name and request.form[
-                'description'] == edited_item.description and request.form[
-                'image'] == edited_item.image:
-            return redirect(url_for('show_category', category_id=category_id))
+    if 'username' in login_session:
+        category = session.query(Category).filter_by(id=category_id).one()
+        edited_item = session.query(Item).filter_by(id=item_id).one()
+        if request.method == 'POST':
+            if request.form['name'] == edited_item.name and request.form[
+                    'description'] == edited_item.description and request.form[
+                    'image'] == edited_item.image:
+                return redirect(url_for('show_category', category_id=category_id))
+            else:
+                if request.form['name']:
+                    edited_item.name = request.form['name']
+                if request.form['description']:
+                    edited_item.description = request.form['description']
+                if request.form['image']:
+                    edited_item.image = request.form['image']
+                session.add(edited_item)
+                session.commit()
+                flash('Item %s edited' % edited_item.name)
+                return redirect(url_for('show_category', category_id=category_id))
         else:
-            if request.form['name']:
-                edited_item.name = request.form['name']
-            if request.form['description']:
-                edited_item.description = request.form['description']
-            if request.form['image']:
-                edited_item.image = request.form['image']
-            session.add(edited_item)
-            session.commit()
-            flash('Item %s edited' % edited_item.name)
-            return redirect(url_for('show_category', category_id=category_id))
+            return render_template('editItem.html', category=category, item=edited_item)
     else:
-        return render_template('editItem.html', category=category, item=edited_item)
+        return redirect(url_for('login_page'))
 
 
 # delete an item
 @app.route("/categories/<int:category_id>/<int:item_id>/delete/", methods=['GET', 'POST'])
 def delete_item(category_id, item_id):
-    category = session.query(Category).filter_by(id=category_id).one()
-    deleted_item = session.query(Item).filter_by(id=item_id).one()
-    if request.method == 'POST':
-        session.delete(deleted_item)
-        return redirect(url_for('show_category', category_id=category_id))
+    if 'username' in login_session:
+        category = session.query(Category).filter_by(id=category_id).one()
+        deleted_item = session.query(Item).filter_by(id=item_id).one()
+        if request.method == 'POST':
+            session.delete(deleted_item)
+            return redirect(url_for('show_category', category_id=category_id))
+        else:
+            return render_template('deleteItem.html', category=category, item=deleted_item)
     else:
-        return render_template('deleteItem.html', category=category, item=deleted_item)
+        return redirect(url_for('login_page'))
 
 
 # show item page
@@ -151,7 +172,8 @@ def delete_item(category_id, item_id):
 def show_item(category_id, item_id):
     category = session.query(Category).filter_by(id=category_id).one()
     item = session.query(Item).filter_by(id=item_id).one()
-    return render_template('item.html', category=category, item=item)
+    loggedin = 'username' in login_session
+    return render_template('item.html', category=category, item=item, loggedin=loggedin)
 
 
 @app.route("/categories/<int:category_id>/<int:item_id>/JSON/")
@@ -253,19 +275,14 @@ def get_user_id(email):
     try:
         user = session.query(User).filter_by(email=email).one()
         return user.id
-    except:
+    except Exception as e:
+        print e
         return None
 
 
-@app.route('/disconnect')
-def disconnect():
+@app.route('/fbdisconnect', methods=['POST'])
+def fbdisconnect():
     if 'provider' in login_session:
-        facebook_id = login_session['facebook_id']
-        access_token = login_session['access_token']
-        url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id, access_token)
-        h = httplib2.Http()
-        result = h.request(url, 'DELETE')[1]
-        print result
         del login_session['facebook_id']
         del login_session['username']
         del login_session['email']
